backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy>=2.0.0
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.18
pydantic==2.5.3
pydantic-settings==2.1.0
ortools==9.8.3296
pandas==2.1.4
openpyxl==3.1.2
python-dateutil==2.8.2
email-validator==2.1.0
reportlab==4.0.9
python-docx==1.1.0
